+++ src/seagulls/engine/_game_clock.py
@@ -1,7 +1,7 @@
 from pygame.time import Clock
 
 from ._game_object import GameObject
-from ._pyagme import Surface
+from ._pygame import Surface
 
 
 class GameClock(GameObject):
