+++ src/seagulls/examples/space_shooter/__init__.py
@@ -1,14 +1,12 @@
+from ._active_ship_client import ActiveShipClient
 from ._asteroid_field import AsteroidField
-from ._selectable_ship_menu import (
-    ActiveShipClient,
-    BlueShip,
-    IShip,
-    OrangeShip,
-    ShipCatalog,
-    ShipSelectionMenu
-)
+from ._blue_ship import BlueShip
+from ._orange_ship import OrangeShip
+from ._score_tracker import ScoreTracker
+from ._selectable_ship_menu import ShipCatalog, ShipSelectionMenu
 from ._ship import Ship
-from ._shooter_scene import ScoreOverlay, ScoreTracker, ShooterScene
+from ._ship_interfaces import IShip
+from ._shooter_scene import ScoreOverlay, ShooterScene
 from ._space_collisions import SpaceCollisions
 
 __all__ = [
