+++ src/seagulls/examples/space_shooter/_check_game_rules_interface.py
@@ -1,0 +1,8 @@
+from abc import ABC, abstractmethod
+
+
+class ICheckGameRules(ABC):
+
+    @abstractmethod
+    def check(self) -> None:
+        pass
